move icon to /icons
