+++ airbyte-webapp/src/components/icons/ModificationIcon.tsx
@@ -1,6 +1,6 @@
 export const ModificationIcon: React.FC = () => {
   return (
-    <svg width="20" height="20" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <svg width="20" height="20" viewBox="0 0 20 20" fill="none">
       <path
         fillRule="evenodd"
         clipRule="evenodd"
